[chore] [pdata] Split Value/Map/Slice into separate files
